Add Mason and setup cursor line option
This commit adds the `mason-org/mason.nvim` plugin to the configuration and sets the `cursorline` option to improve the visual experience in Vim. Additionally, it introduces a new keymap `<leader>lT` to toggle all LSP servers, enhancing the developer's workflow by providing a quick way to stop or start all language server processes.